Move test to tests dir
