+++ tests/__init__.py
@@ -1,0 +1,5 @@
+from . import test_align
+from . import test_match
+from . import test_midi
+from . import test_postalign
+from . import test_processfile
